+++ src/ecc/operations.ts
@@ -1,6 +1,7 @@
+import aes from '../aes'
 import utils from '../utils'
 import { ECC_READ_ALG, ECC_WRITE_ALG, DEFAULT_SYMM_ALG, DEFAULT_SYMM_LEN } from '../constants'
-import { PrivateKey, PublicKey, HashAlg, SymmKey, SymmKeyOpts, SymmAlg, CipherText } from '../types'
+import { PrivateKey, PublicKey, HashAlg, SymmKey, SymmKeyOpts, CipherText } from '../types'
 
 export async function signBytes(data: ArrayBuffer, privKey: PrivateKey, hashAlg: HashAlg): Promise<ArrayBuffer> {
   return window.crypto.subtle.sign(
@@ -32,48 +33,14 @@ export async function getSharedKey(privateKey: PrivateKey, publicKey: PublicKey,
   )
 }
 
-export async function symmEncrypt(data: ArrayBuffer, key: SymmKey, opts?: Partial<SymmKeyOpts>): Promise<CipherText> {
-  const alg = opts?.alg || DEFAULT_SYMM_ALG
-  const iv = opts?.iv || utils.randomBuf(16)
-  const cipherBuf = await window.crypto.subtle.encrypt(
-    { 
-      name: alg,
-      length: opts?.length || DEFAULT_SYMM_LEN,
-      // AES-CTR uses a counter, AES-GCM/AES-CBC use an initialization vector
-      iv: alg === SymmAlg.AES_CTR ? undefined : iv,
-      counter: alg === SymmAlg.AES_CTR ? new Uint8Array(iv) : undefined,
-    },
-    key,
-    data
-  )
-  return utils.joinBufs(iv, cipherBuf)
-}
-
-export async function symmDecrypt(cipherText: CipherText, key: SymmKey, opts?: Partial<SymmKeyOpts>): Promise<ArrayBuffer> {
-  const alg = opts?.alg || DEFAULT_SYMM_ALG
-  const iv = cipherText.slice(0, 16)
-  const cipherBytes = cipherText.slice(16)
-  const msgBuff = await window.crypto.subtle.decrypt(
-    { name: alg,
-      length: opts?.length || DEFAULT_SYMM_LEN,
-      // AES-CTR uses a counter, AES-GCM/AES-CBC use an initialization vector
-      iv: alg === SymmAlg.AES_CTR ? undefined : iv,
-      counter: alg === SymmAlg.AES_CTR ? new Uint8Array(iv) : undefined,
-    },
-    key,
-    cipherBytes
-  )
-  return msgBuff
-}
-
 export async function encryptBytes(data: ArrayBuffer, privateKey: PrivateKey, publicKey: PublicKey, opts?: Partial<SymmKeyOpts>): Promise<CipherText> {
   const cipherKey = await getSharedKey(privateKey, publicKey, opts)
-  return symmEncrypt(data, cipherKey, opts)
+  return aes.encrypt(data, cipherKey, opts)
 }
 
 export async function decryptBytes(cipherText: CipherText, privateKey: PrivateKey, publicKey: PublicKey, opts?: Partial<SymmKeyOpts>): Promise<ArrayBuffer> {
   const cipherKey = await getSharedKey(privateKey, publicKey, opts)
-  return symmDecrypt(cipherText, cipherKey, opts)
+  return aes.decrypt(cipherText, cipherKey, opts)
 }
 
 export async function getPublicKey(keypair: CryptoKeyPair): Promise<string> {
@@ -85,8 +52,6 @@ export default {
   signBytes,
   verifyBytes,
   getSharedKey,
-  symmDecrypt,
-  symmEncrypt,
   encryptBytes,
   decryptBytes,
   getPublicKey
